+++ client/FullSize/NameCollector/NameCollector.js
@@ -1,0 +1,36 @@
+import FullSize from "../FullSize";
+import DrawUtil from "../../Draw/Draw";
+
+export default class NameCollector extends FullSize {
+    constructor(_screen){
+        super(_screen);
+        this.name ='';
+        this.dirty = false;
+        this.render = function () {
+            let vm = this;
+            let yOffset = 580;
+            DrawUtil.square(vm.cX - 200, yOffset - 235, 400, 40, '#208C30', 1);
+            DrawUtil.square(vm.cX - 195, yOffset - 230, 390, 30, 'black', 1);
+            if(vm.timer < 30) DrawUtil.text('>',vm.cX - 190, yOffset - 205, '#208C30', 24);
+            if(!this.dirty) DrawUtil.text('Enter Name',vm.cX - 170, yOffset - 207, '#208C30', 24);
+            if(this.dirty) DrawUtil.text(this.name,vm.cX - 170, yOffset - 207, '#208C30', 24);
+        };
+
+        this.onAnyKeyDown = function (key) {
+
+            if(this.name.length > 10 && key !== 'Backspace') return;
+
+            if (key === 'Backspace' && this.name !== '') {
+                this.name = this.name.substring(0, this.name.length - 1);
+            } else {
+                if(key.length === 1) {
+                    this.dirty = true;
+                    this.name += key;
+                }
+            }
+
+            socket.emit('update-name', this.name)
+
+        };
+    }
+}
