+++ client/FullSize/TitleCard/TitleCard.js
@@ -1,0 +1,39 @@
+import FullSize from "../FullSize";
+import DrawUtil from "../../Draw/Draw";
+
+export default class TitleCard extends FullSize {
+    constructor(_screen) {
+        super(_screen);
+        let vm = this;
+        this.render = function () {
+
+            let linkOffset = 280;
+            let carrotOffset = 440;
+
+            let ctx = DrawUtil.getCtx();
+            //Text LINK
+
+            ctx.globalAlpha = 0.6;
+            ctx.font = "240px Arial Black";
+            ctx.fillStyle = '#083F10';
+            if (this.timer <= 30) ctx.fillText(">", (this.width / 2) + 5 - carrotOffset, this.y + 5 + 275);
+            ctx.fillText("LINK", (this.width / 2) - 5 - linkOffset, this.y + 5 + 275);
+            ctx.fillStyle = '#208C30';
+            if (this.timer >= 30) ctx.globalAlpha = 0.2;
+            ctx.fillText(">", (this.width / 2) - carrotOffset, this.y + 275);
+            if (this.timer >= 30) ctx.globalAlpha = 0.6;
+            ctx.fillText("LINK", (this.width / 2) - linkOffset, this.y + 275);
+
+            ctx.font = "24px Arial Black";
+            for (let i = 0; i < 36; i++) {
+                [80, 335].forEach(function (y) {
+                    ctx.fillText('= ', (i * 25) + (vm.width / 2) - 455, y);
+                });
+                [430, -460].forEach(function (x) {
+                    ctx.fillText('+', x + (vm.width / 2), (i * 7) + 88)
+                })
+            }
+            ctx.globalAlpha = 1;
+        }
+    }
+}
