+++ package.json
@@ -17,10 +17,11 @@
     "node": "*"
   },
   "dependencies": {
-    "@babel/core": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
+    "@babel/core": "^7.1.5",
+    "@babel/preset-env": "^7.1.5",
     "archiver": "^2.1.1",
     "babel": "^6.23.0",
+    "babel-loader": "^8.0.4",
     "babel-preset-es2015": "^6.24.1",
     "body-parser": "^1.18.3",
     "del": "^3.0.0",
@@ -36,7 +37,10 @@
     "sequelize": "^4.38.0",
     "socket.io": "^2.1.0",
     "sqlite3": "^4.0.1",
-    "uglify-js": "^3.4.9"
+    "uglify-js": "^3.4.9",
+    "underscore": "^1.9.1",
+    "webpack": "^4.25.1",
+    "webpack-cli": "^3.1.2"
   },
   "devDependencies": {
     "mocha": "^5.2.0"
