+++ public/index.html
@@ -11,8 +11,7 @@
 <canvas id="a" style="height: 100%; width: 100%; position: absolute"></canvas>
 
 <script src="/socket.io/socket.io.js"></script>
-<script src="/shared.js"></script>
-<script src="/client.js"></script>
+<script src="/bundle.js"></script>
 
 </body>
 </html>
