+++ shared/Entity/Projectile/PistolProjectile/PistolProjectile.js
@@ -1,0 +1,7 @@
+import Projectile from '../Projectile'
+
+export default class PistolProjectile extends Projectile {
+    constructor(nonce, x, y, rotationDegrees, fireTime, playerNonce) {
+        super(nonce, x, y, rotationDegrees, fireTime, playerNonce, 1, 8, 9, 10,  4 * 1000, 5);
+    }
+}
