+++ shared/Entity/Projectile/ShotgunProjectile/ShotgunProjectile.js
@@ -1,0 +1,7 @@
+import Projectile from '/shared/Entity/Projectile/Projectile'
+
+export default class ShotgunProjectile extends Projectile {
+    constructor(nonce, x, y, rotationDegrees, fireTime, playerNonce) {
+        super(nonce, x, y, rotationDegrees, fireTime, playerNonce, 9, 5, 8, 50, 900, 1);
+    }
+}
