+++ shared/MiscUtil/MiscUtil.js
@@ -1,0 +1,20 @@
+//TODO: This is just a grab bag of random stuff that should eventually be better organized...
+
+export default class MiscUtil {
+
+    static getStatusVals(winStatus) {
+        switch (winStatus) {
+            case 'TIE':
+                return {color: '#cd7f32', text: 'LINK TIE'};
+            case 'WINNER':
+                return {color: 'gold', text: 'VICTORY'};
+            default:
+                return {color: 'red', text: 'TERMINATED'};
+        }
+    }
+
+    static mouseInBounds(x, y, height, width) {
+        return mousePos.x > x && mousePos.x < x + width && mousePos.y > y && mousePos.y < y + height;
+    }
+
+}
